moved gear code to gear component
